refactor: move nitro testnode networks to new file
