+++ packages/arb-token-bridge-ui/tests/support/common.ts
@@ -10,8 +10,7 @@ import {
   defaultL2Network,
   defaultL3Network,
   defaultL3CustomGasTokenNetwork
-} from '../../src/util/networks'
-import { getChainIdFromProvider } from '../../src/token-bridge-sdk/utils'
+} from '../../src/util/networksNitroTestnode'
 
 export type NetworkType = 'parentChain' | 'childChain'
 export type NetworkName =
@@ -276,7 +275,7 @@ export async function checkForAssertions({
 
   const rollupContract = new ethers.Contract(rollupAddress, abi, parentProvider)
 
-  const parentChainId = await getChainIdFromProvider(parentProvider)
+  const parentChainId = (await parentProvider.getNetwork()).chainId
 
   try {
     while (true) {
